Backend/requirements.txt
fastapi
uvicorn
firebase-admin
google-cloud-firestore
google-cloud-storage
openai-whisper
websockets
scipy
numpy
pydantic-settings
certifi
python-multipart